requirements.txt
streamlit==1.28.1
pypdf==5.0.0
pdf2image==1.17.0
Pillow==10.4.0
python-pptx==0.6.23
pdf2docx==0.5.8
tabula-py==2.9.3
pdfminer.six==20231228
tqdm==4.66.5
pandas==2.2.2
openpyxl==3.1.5
reportlab==4.2.2
